BotsProjects/yapb/ai_training/requirements.txt
# YaPB Neural Zombie AI Requirements
# Install with: pip install -r requirements.txt

# Core machine learning frameworks
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Data processing and analysis
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Logging and monitoring
tensorboard>=2.7.0
wandb>=0.12.0  # Optional: for cloud logging

# Neural network utilities
onnx>=1.10.0  # For model export
onnxruntime>=1.9.0  # For inference optimization

# Data serialization and I/O
h5py>=3.4.0
joblib>=1.1.0

# Configuration and utilities
pyyaml>=5.4.0
tqdm>=4.62.0
psutil>=5.8.0  # For system monitoring

# Optional: GPU acceleration
# Install separately if you have CUDA:
# torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# Optional: Advanced neural network libraries
transformers>=4.12.0  # For attention mechanisms
stable-baselines3>=1.3.0  # For RL algorithms

# Windows-specific dependencies for memory reading and screen capture
pymem>=1.8.5
pywin32>=306
opencv-python>=4.6.0
gymnasium>=0.26.0
pillow>=8.3.0
